changes in user controller functions
